Update example Terminal profile
The previous profile had colors with round numbers (like "FFFFFF" ->
1,1,1 or "00FF00" -> 0,1,0), which get converted to integers by
Terminal.app's profile export.

Converting the numbers to intergers in Colorscheme would add a lot of
complexity, so we'll treat every color it gets as a Float instead.

To be able to test this with a profile exported from Terminal.app, the
test/example.terminal profile was updated so each color is a float (red,
which was "FF0000" is now "FF0001", for example), ignoring the round
numbers case.